svg/painting/parsing/fill-opacity-valid.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="800" height="800"><title>SVG Painting: parsing fill-opacity with valid values</title><metadata><h:link href="https://svgwg.org/svg2-draft/painting.html#FillOpacityProperty" rel="help"/><h:meta content="fill-opacity supports the full grammar 'alpha-value'." name="assert"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/parsing-testcommon.js"/><script><![CDATA[

test_valid_value("fill-opacity", "-1");
test_valid_value("fill-opacity", "0.5");
test_valid_value("fill-opacity", "3");
test_valid_value("fill-opacity", "-100%", "-1");
test_valid_value("fill-opacity", "50%", "0.5");
test_valid_value("fill-opacity", "300%", "3");

  ]]></script></svg>
